Refactored additional components that I feel need some extra little ones
