+++ src/components/consumer-pro-protection/hero-texts.js
@@ -1,0 +1,23 @@
+import React from "react";
+
+export const HeroTexts = () => {
+  return (
+    <div className="mt-12 mb-6 md:mx-auto max-w-xl">
+      <h2 className="font-semibold inline-flex mt-2 text-sm tracking-wide uppercase bg-gradient-to-r from-teal-500 to-cyan-800 bg-clip-text text-transparent">
+        Big or small projects, we can help
+      </h2>
+
+      <div className="my-12">
+        <h2 className="mt-4 text-4xl font-extrabold tracking-tight text-blueGray-800 flex items-center">
+          How we protect you and our professionals
+        </h2>
+
+        <p className="mt-3 font-medium text-base text-gray-500 sm:text-lg md:mt-5 md:max-w-3xl">
+          We have a two-way mission. Offering our customers the right Pro for
+          their needs, and assisting our Pros in gaining real customers. Here
+          are the steps to this process:
+        </p>
+      </div>
+    </div>
+  );
+};
