+++ src/pages/index.js
@@ -1,8 +1,8 @@
 import { CallToAction } from "../components/call-to-action";
 import { FeaturedTestimonial } from "../components/featured-testimonial";
 import { Hero } from "../components/hero-section/hero";
-import { HowItWorks } from "../components/how-it-works";
-import { HowWeProtect } from "../components/how-we-protect";
+import { HowItWorks } from "../components/how-it-works/how-it-works";
+import { HowWeProtect } from "../components/consumer-pro-protection/how-we-protect";
 import Layout from "../components/layout";
 import React from "react";
