+++ src/utils/card-details.js
@@ -1,0 +1,26 @@
+export const cardDetails = [
+  {
+    id: 1,
+    unit: "7K+",
+    description: "Happy Customers",
+    ping_full:
+      "animate-ping absolute inline-flex h-full w-full rounded-full bg-yellow-400 opacity-75",
+    ping_fixed: "relative inline-flex rounded-full h-3 w-3 bg-yellow-500",
+  },
+  {
+    id: 2,
+    unit: "$300-1K+",
+    description: "Cost Savings",
+    ping_full:
+      "animate-ping absolute inline-flex h-full w-full rounded-full bg-rose-400 opacity-75",
+    ping_fixed: "relative inline-flex rounded-full h-3 w-3 bg-rose-500",
+  },
+  {
+    id: 3,
+    unit: "96%",
+    description: "Same Day Repairs",
+    ping_full:
+      "animate-ping absolute inline-flex h-full w-full rounded-full bg-cyan-400 opacity-75",
+    ping_fixed: "relative inline-flex rounded-full h-3 w-3 bg-cyan-500",
+  },
+];
